Passage de la logique d'authentification dans un service et dans les controllers

+++ packages/backend/src/controllers/user.controller.ts
@@ -9,9 +9,9 @@ export const userController = {
       const newUser = { name, email, role, phone, passwordHash, companyId };
 
       const createdUser = await userModel.createUser(newUser);
-      APIResponse(res, { id: createdUser[0].id, ...newUser }, "User created", 201);
+      return APIResponse(res, { id: createdUser[0].id, ...newUser }, "User created", 201);
     } catch (error: any) {
-      APIResponse(res, null, error.message, 400);
+      return APIResponse(res, null, error.message, 400);
     }
   },
 
@@ -21,11 +21,11 @@ export const userController = {
       const user = await userModel.getById(id);
       if (!user) {
         throw new Error("User not found");
-      } else {
-        APIResponse(res, user, "User found", 201);
       }
+
+      return APIResponse(res, user, "User found", 201);
     } catch (error: any) {
-      APIResponse(res, null, error.message, 400);
+      return APIResponse(res, null, error.message, 400);
     }
   },
 
@@ -36,11 +36,11 @@ export const userController = {
       const products = await userModel.getUserProducts(id);
       if (!products) {
         throw new Error("Products not found");
-      } else {
-        APIResponse(res, products, "Products found", 201);
       }
+      
+      return APIResponse(res, products, "Products found", 201);
     } catch (error: any) {
-      APIResponse(res, null, error.message, 400);
+      return APIResponse(res, null, error.message, 400);
     }
   },
 
@@ -48,20 +48,32 @@ export const userController = {
   async update (req: Request, res: Response) {
     try {
       const { id } = req.params;
+
+      const existingUser = await userModel.getById(id);
+      if (!existingUser) {
+        throw new Error("User not found");
+      }
+
       const updatedUser = await userModel.updateById(id, req.body);
-      APIResponse(res, updatedUser, "User updated");
+      return APIResponse(res, updatedUser, "User updated");
     } catch (error: any) {
-      APIResponse(res, null, error.message, 500);
+      return APIResponse(res, null, error.message, 500);
     }
   },
 
   async delete (request: Request, response: Response) {
     try {
       const { id } = request.params;
+
+      const existingUser = await userModel.getById(id);
+      if (!existingUser) {
+        throw new Error("User not found");
+      }
+
       const deletedUser = await userModel.deleteById(id);
-      APIResponse(response, deletedUser, "User deleted");
+      return APIResponse(response, deletedUser, "User deleted");
     } catch (error: any) {
-      APIResponse(response, null, error.message, 500);
+      return APIResponse(response, null, error.message, 500);
     }
   },
 };
